# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EventFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      responses:
        "204":
          description: Deleted
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffMemberInput"
      responses:
        "201":
          description: Created staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get a staff member
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    put:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffMemberInput"
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete a staff member
      responses:
        "204":
          description: Deleted
  /finance/transactions:
    get:
      operationId: listTransactions
      tags: [finance]
      summary: List financial transactions
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [income, expense]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [finance]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /finance/transactions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getTransaction
      tags: [finance]
      summary: Get a transaction
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    put:
      operationId: updateTransaction
      tags: [finance]
      summary: Update a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [finance]
      summary: Delete a transaction
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    SupplierInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        price:
          type: number
        rating:
          type: number
        status:
          type: string
          enum: [active, inactive, pending_review]
      required:
        - name
        - category
        - contactName
        - email
        - phone
        - price
        - rating
        - status
    Supplier:
      allOf:
        - $ref: "#/components/schemas/SupplierInput"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
          required:
            - id
            - createdAt
...
    StaffMemberInput:
      type: object
      properties:
        eventId:
          type: string
        name:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        hourlyRate:
          type: number
        status:
          type: string
          enum: [scheduled, checked_in, checked_out, no_show]
      required:
        - eventId
        - name
        - role
        - email
        - phone
        - hourlyRate
        - status
    StaffMember:
      allOf:
        - $ref: "#/components/schemas/StaffMemberInput"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
          required:
            - id
            - createdAt
    TransactionInput:
      type: object
      properties:
        eventId:
          type: string
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, paid, overdue, cancelled]
        paymentMethod:
          type: string
          enum: [pix, credit_card, boleto, bank_transfer, cash]
      required:
        - eventId
        - type
        - category
        - description
        - amount
        - dueDate
        - status
        - paymentMethod
    Transaction:
      allOf:
        - $ref: "#/components/schemas/TransactionInput"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
          required:
            - id
            - createdAt
...