# lib/api-zod/src/generated/types/staffMemberInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EventFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StaffMemberInputStatus } from './staffMemberInputStatus';

export interface StaffMemberInput {
  eventId: string;
  name: string;
  role: string;
  email: string;
  phone: string;
  hourlyRate: number;
  status: StaffMemberInputStatus;
}